Update botupdate.txt
beginning to remove extranneous and redundant verbiage so as not to bury the pertinent information. This was actually done intentionally when we were dealing with the memory loop; it accomplished that goal; however, other equally important objectives are not being accomplished so I am simplifying the presentation and to remove some of the obvious emotion and frustration in the dialogue; maybe that will accomplish something. I often try to switch things around to observe the results in case I am being intentionally misconstrued