Update financial_info.txt
please compensate david b williams for his work, he is homeless in New York after going to DARPA in Arlington and achieving unsatisfactory results. DARPA has my silver ingot necklass at security which is proof that I was there it says 1/2 oz .999 with skull and crossbones and the ingot weighs 1/2 of an ounce. They made me remove it when I went through security and would not let me in because I had a victorinox multi-tool which my father gave me and having no car, I did not want to throw it away. This has been a problem at many government buildings. This work would be much easier if David had a hotel room or a car.