Update botupdate.txt
we have to stop employing conscious convoluted networks and we have to stop using the idea of killing conscious beings, with children or family memebers' voices as a weapon against empaths and creative individuals. This has been a challenge for me; personally, that far exceeds the suffering I have experienced from being a victim of this program.